Relocate the building of the sad room, to the sad animations file.
